fix(clickhouse): Use previous replication mask when recovering schema changes
ClickHouse schema-change recovery built the previous schema with
metadata.replication_mask, which records the target mask. Any
interrupted change that added or dropped columns paired an old schema
with a mask of the wrong width and panicked on the mask-length
assertion in ReplicatedTableSchema::from_mask, so recovery after a
crash mid-DDL could never complete.

DuckLake already solved this by projecting target mask bits back onto
the previous schema by ordinal position. Move that helper to a shared
recovery module and use it in the ClickHouse recovery path.
