+++ crates/etl-destinations/src/lib.rs
@@ -4,6 +4,8 @@
 //! warehouses and analytics platforms, enabling data replication from Postgres
 //! to cloud services.
 
+#[cfg(any(feature = "clickhouse", feature = "ducklake"))]
+mod recovery;
 #[cfg(any(feature = "bigquery", feature = "ducklake", feature = "snowflake"))]
 mod retry;
 #[cfg(any(feature = "ducklake", feature = "snowflake"))]
